file renames + add comment for clarity that root is never modified
