+++ src/viam/ur/module/ur_arm_model.cpp
@@ -1,4 +1,4 @@
-#include "ur_models.hpp"
+#include "ur_arm_model.hpp"
 
 #include <array>
 #include <cmath>
@@ -10,12 +10,15 @@
 #include <string>
 #include <type_traits>
 #include <unordered_map>
+#include <utility>
+#include <vector>
 
 #include <Eigen/Geometry>
 
 #include <json/json.h>
 
 #include "rust_utils.hpp"
+#include "ur_arm.hpp"
 #include "utils.hpp"  // degrees_to_radians / radians_to_degrees
 
 namespace {
@@ -323,6 +326,46 @@ viam::sdk::pose apply_correction_to_pose(const viam::sdk::pose& p, const Eigen::
 
 }  // namespace
 
+const std::vector<UrModelDescriptor>& UrModelDescriptor::all() {
+    static const std::vector<UrModelDescriptor> table = [] {
+        std::vector<UrModelDescriptor> t;
+        // ur3e/ur7e have no shipped meshes; `get_3d_models` returns an
+        // empty map for them.
+        t.push_back({"ur3e", "ur3", {}});
+        t.push_back(
+            {"ur5e", "ur5", {"base_link", "ee_link", "shoulder_link", "forearm_link", "upper_arm_link", "wrist_1_link", "wrist_2_link"}});
+        t.push_back({"ur7e", "ur5", {}});
+        t.push_back({"ur20",
+                     "ur20",
+                     {"base_link", "wrist_3_link", "shoulder_link", "forearm_link", "upper_arm_link", "wrist_1_link", "wrist_2_link"}});
+        return t;
+    }();
+    return table;
+}
+
+const UrModelDescriptor& UrModelDescriptor::for_sdk_name(const std::string& sdk_name) {
+    for (const auto& d : all()) {
+        if (d.sdk_name == sdk_name) {
+            return d;
+        }
+    }
+    throw std::invalid_argument("UrModelDescriptor::for_sdk_name: unknown sdk model name `" + sdk_name + "`");
+}
+
+UrArmModel::UrArmModel(viam::sdk::Model sdk_model, const UrModelDescriptor& descriptor)
+    : sdk_model_(std::move(sdk_model)), descriptor_(&descriptor) {}
+
+UrArmModel::UrArmModel(viam::sdk::Model sdk_model)
+    // sdk_model_ is initialized in declaration order before descriptor_,
+    // so reading the moved-into sdk_model_ for the descriptor lookup here
+    // is well-defined (a delegating-constructor form risks reading from
+    // the moved-from parameter, depending on argument evaluation order).
+    : sdk_model_(std::move(sdk_model)), descriptor_(&UrModelDescriptor::for_sdk_name(sdk_model_.model_name())) {}
+
+UrArmModel UrArmModel::from_sdk_name(const std::string& sdk_name) {
+    return UrArmModel{URArm::model(sdk_name), UrModelDescriptor::for_sdk_name(sdk_name)};
+}
+
 UrArmModel::Kinematics::Kinematics(UrArmModel m) : model(std::move(m)), limits{}, link_locals{}, geometries{}, link_names{} {
     for (auto& mat : link_locals) {
         mat = Eigen::Matrix4d::Identity();
@@ -391,6 +434,12 @@ UrArmModel::Kinematics UrArmModel::load_kinematics(const std::filesystem::path& 
     // from "world" without ever looking up entries by name. Per the shipped
     // JSONs each parent has at most one child (no branching); we enforce
     // that here.
+    //
+    // Invariant: `child_of` stores raw pointers into `root`, so `root` must
+    // not be mutated while `child_of` is in scope or those pointers may
+    // dangle. The chain walk below honors this by only reading existing
+    // members (`entry["id"]`, `entry.isMember(...)`) -- never assigning into
+    // `root` or its children.
     std::unordered_map<std::string, const Json::Value*> child_of;
     auto index_entry = [&](const Json::Value& entry, const char* kind) {
         if (!entry.isMember("id") || !entry["id"].isString()) {
